+++ src/main/java/com/kestrel/weddingbookkeeper/api/member/dao/MemberDao.java
@@ -1,14 +1,13 @@
 package com.kestrel.weddingbookkeeper.api.member.dao;
 
 import com.kestrel.weddingbookkeeper.api.member.vo.Member;
-import java.util.List;
 import java.util.Optional;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface MemberDao {
 
-    Optional<Member> selectById(Integer id);
+    Optional<Member> selectById(Long id);
 
     int insertMember(Member member);
