+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/dto/MemberWeddingSaveDto.java
@@ -4,15 +4,15 @@ import com.kestrel.weddingbookkeeper.api.member.vo.Member;
 
 public class MemberWeddingSaveDto {
 
-    private Integer memberId;
-    private Integer weddingId;
+    private Long memberId;
+    private Long weddingId;
     private int donationAmount;
     private String relation;
     private Boolean isGroomSide;
     private String guestName;
     private String fcmToken;
 
-    public MemberWeddingSaveDto(Member member, Integer weddingId, MemberWeddingDto memberWeddingDto) {
+    public MemberWeddingSaveDto(Member member, Long weddingId, MemberWeddingDto memberWeddingDto) {
         this.memberId = member.getMemberId();
         this.weddingId = weddingId;
         this.donationAmount = memberWeddingDto.getDonationAmount();
