+++ src/main/java/com/kestrel/weddingbookkeeper/api/wedding/dto/WeddingUpdateDto.java
@@ -2,15 +2,15 @@ package com.kestrel.weddingbookkeeper.api.wedding.dto;
 
 public class WeddingUpdateDto {
 
-    private Integer weddingId;
+    private Long weddingId;
     private String qrImgUrl;
 
-    public WeddingUpdateDto(Integer weddingId, String qrImgUrl) {
+    public WeddingUpdateDto(Long weddingId, String qrImgUrl) {
         this.weddingId = weddingId;
         this.qrImgUrl = qrImgUrl;
     }
 
-    public Integer getWeddingId() {
+    public Long getWeddingId() {
         return weddingId;
     }
